+++ products/future.go
@@ -44,14 +44,15 @@ type Future struct {
 	settlementPriceListener    func(context.Context, *num.Uint)
 }
 
-func (f *Future) Unsubscribe(ctx context.Context, oe OracleEngine) {
-	oe.Unsubscribe(ctx, f.oracle.settlementPriceSubscriptionID)
-	oe.Unsubscribe(ctx, f.oracle.tradingTerminatedSubscriptionID)
+func (f *Future) Unsubscribe(ctx context.Context) {
+	f.oracle.unsubscribe(ctx, f.oracle.settlementPriceSubscriptionID)
+	f.oracle.unsubscribe(ctx, f.oracle.tradingTerminatedSubscriptionID)
 }
 
 type oracle struct {
 	settlementPriceSubscriptionID   oracles.SubscriptionID
 	tradingTerminatedSubscriptionID oracles.SubscriptionID
+	unsubscribe                     oracles.Unsubscriber
 	binding                         oracleBinding
 	data                            oracleData
 	settlementPriceDecimals         uint32
@@ -246,7 +247,7 @@ func NewFuture(ctx context.Context, log *logging.Logger, f *types.Future, oe Ora
 		return nil, fmt.Errorf("invalid oracle spec binding for settlement price: %w", err)
 	}
 
-	future.oracle.settlementPriceSubscriptionID = oe.Subscribe(ctx, *oracleSpecForSettlementPrice, future.updateSettlementPrice)
+	future.oracle.settlementPriceSubscriptionID, future.oracle.unsubscribe = oe.Subscribe(ctx, *oracleSpecForSettlementPrice, future.updateSettlementPrice)
 
 	if log.IsDebug() {
 		log.Debug("future subscribed to oracle engine for settlement price",
@@ -277,7 +278,7 @@ func NewFuture(ctx context.Context, log *logging.Logger, f *types.Future, oe Ora
 		return nil, fmt.Errorf("invalid oracle spec binding for trading termination: %w", err)
 	}
 
-	future.oracle.tradingTerminatedSubscriptionID = oe.Subscribe(ctx, *oracleSpecForTerminatedMarket, tradingTerminationCb)
+	future.oracle.tradingTerminatedSubscriptionID, _ = oe.Subscribe(ctx, *oracleSpecForTerminatedMarket, tradingTerminationCb)
 
 	if log.IsDebug() {
 		log.Debug("future subscribed to oracle engine for market termination event",